More Professional Update to README.md
This is an update to the README to make it more clear as to what this project is about and how others can contribute to it. It also functions to make the README more professional as well as remove the first person tense from the README (i.e. "We") and corrects the grammar.